Split global promos to it's own repository and rename PromoRepository to PromoHomepageRepository so it encourages developers to make a new file and not put it all in one.
