+++ app/Repositories/PromoGlobalRepository.php
@@ -3,12 +3,11 @@
 namespace App\Repositories;
 
 use Contracts\Repositories\DataRepositoryContract;
-use Contracts\Repositories\PromoRepositoryContract;
 use Illuminate\Cache\Repository;
 use Waynestate\Api\Connector;
 use Waynestate\Promotions\ParsePromos;
 
-class PromoRepository implements DataRepositoryContract, PromoRepositoryContract
+class PromoGlobalRepository implements DataRepositoryContract
 {
     /** @var Connector */
     protected $wsuApi;
@@ -133,33 +132,5 @@ class PromoRepository implements DataRepositoryContract, PromoRepositoryContract
         unset($promos['main_contact']);
 
         return $promos;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getHomepagePromos()
-    {
-        $group_reference = [
-            123 => 'example',
-        ];
-
-        $group_config = [
-            'example' => 'first',
-        ];
-
-        $params = [
-            'method' => 'cms.promotions.listing',
-            'promo_group_id' => array_keys($group_reference),
-            'filename_url' => true,
-            'is_active' => '1',
-        ];
-
-        $promos = $this->cache->remember($params['method'].md5(serialize($params)), config('cache.ttl'), function () use ($params) {
-            return $this->wsuApi->sendRequest($params['method'], $params);
-        });
-
-        // Return the parsed promotions
-        return $this->parsePromos->parse($promos, $group_reference, $group_config);
     }
 }
