+++ app/Repositories/PromoHomepageRepository.php
@@ -1,0 +1,62 @@
+<?php
+
+namespace App\Repositories;
+
+use Contracts\Repositories\PromoHomepageRepositoryContract;
+use Illuminate\Cache\Repository;
+use Waynestate\Api\Connector;
+use Waynestate\Promotions\ParsePromos;
+
+class PromoHomepageRepository implements PromoHomepageRepositoryContract
+{
+    /** @var Connector */
+    protected $wsuApi;
+
+    /** @var ParsePromos */
+    protected $parsePromos;
+
+    /** @var Repository */
+    protected $cache;
+
+    /**
+     * Construct the repository.
+     *
+     * @param Connector $wsuApi
+     * @param ParsePromos $parsePromos
+     * @param Repository $cache
+     */
+    public function __construct(Connector $wsuApi, ParsePromos $parsePromos, Repository $cache)
+    {
+        $this->wsuApi = $wsuApi;
+        $this->parsePromos = $parsePromos;
+        $this->cache = $cache;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function get()
+    {
+        $group_reference = [
+            123 => 'example',
+        ];
+
+        $group_config = [
+            'example' => 'first',
+        ];
+
+        $params = [
+            'method' => 'cms.promotions.listing',
+            'promo_group_id' => array_keys($group_reference),
+            'filename_url' => true,
+            'is_active' => '1',
+        ];
+
+        $promos = $this->cache->remember($params['method'].md5(serialize($params)), config('cache.ttl'), function () use ($params) {
+            return $this->wsuApi->sendRequest($params['method'], $params);
+        });
+
+        // Return the parsed promotions
+        return $this->parsePromos->parse($promos, $group_reference, $group_config);
+    }
+}
