+++ styleguide/Repositories/PromoHomepageRepository.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Styleguide\Repositories;
+
+use App\Repositories\PromoHomepageRepository as Repository;
+
+class PromoHomepageRepository extends Repository
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function get()
+    {
+        return [
+            // Example Group
+            //'key' => app('Factories\YourFactory')->create(5),
+        ];
+    }
+}
