+++ tests/Unit/Repositories/PromoGlobalRepositoryTest.php
@@ -5,11 +5,11 @@ namespace Tests\App\Repositories;
 use Tests\TestCase;
 use Mockery as Mockery;
 
-class PromoRepositoryTest extends TestCase
+class PromoGlobalRepositoryTest extends TestCase
 {
     /**
-     * @covers App\Repositories\PromoRepository::__construct
-     * @covers App\Repositories\PromoRepository::getRequestData
+     * @covers App\Repositories\PromoGlobalRepository::__construct
+     * @covers App\Repositories\PromoGlobalRepository::getRequestData
      * @test
      */
     public function getting_promos_with_custom_page_accordion_should_return_accordion_page()
@@ -37,14 +37,14 @@ class PromoRepositoryTest extends TestCase
         $wsuApi->shouldReceive('sendRequest')->with('cms.promotions.listing', Mockery::type('array'))->once()->andReturn($return);
 
         // Get the promos
-        $promos = app('App\Repositories\PromoRepository', ['wsuApi' => $wsuApi])->getRequestData($data);
+        $promos = app('App\Repositories\PromoGlobalRepository', ['wsuApi' => $wsuApi])->getRequestData($data);
 
         $this->assertArrayHasKey('accordion_page', $promos);
     }
 
     /**
-     * @covers App\Repositories\PromoRepository::__construct
-     * @covers App\Repositories\PromoRepository::getRequestData
+     * @covers App\Repositories\PromoGlobalRepository::__construct
+     * @covers App\Repositories\PromoGlobalRepository::getRequestData
      * @test
      */
     public function subsite_using_main_config()
@@ -103,7 +103,7 @@ class PromoRepositoryTest extends TestCase
         $wsuApi->shouldReceive('sendRequest')->with('cms.promotions.listing', Mockery::type('array'))->once()->andReturn($return);
 
         // Get the promos
-        $promos = app('App\Repositories\PromoRepository', ['wsuApi' => $wsuApi])->getRequestData($data);
+        $promos = app('App\Repositories\PromoGlobalRepository', ['wsuApi' => $wsuApi])->getRequestData($data);
 
         // Make sure that the count is equal to the main config limit
         $this->assertCount(1, $promos['contact']);
@@ -111,8 +111,8 @@ class PromoRepositoryTest extends TestCase
     }
 
     /**
-     * @covers App\Repositories\PromoRepository::__construct
-     * @covers App\Repositories\PromoRepository::getRequestData
+     * @covers App\Repositories\PromoGlobalRepository::__construct
+     * @covers App\Repositories\PromoGlobalRepository::getRequestData
      * @test
      */
     public function subsite_using_main_contact_footer()
@@ -159,29 +159,8 @@ class PromoRepositoryTest extends TestCase
         $wsuApi->shouldReceive('sendRequest')->with('cms.promotions.listing', Mockery::type('array'))->once()->andReturn($return);
 
         // Get the promos
-        $promos = app('App\Repositories\PromoRepository', ['wsuApi' => $wsuApi])->getRequestData($data);
+        $promos = app('App\Repositories\PromoGlobalRepository', ['wsuApi' => $wsuApi])->getRequestData($data);
 
         $this->assertCount(1, $promos['contact']);
-    }
-
-    /**
-     * @covers App\Repositories\PromoRepository::getHomepagePromos
-     * @test
-     */
-    public function getting_homepage_promos_should_return_array()
-    {
-        // Fake return
-        $return = [
-            'promotions' => [],
-        ];
-
-        // Mock the connector and set the return
-        $wsuApi = Mockery::mock('Waynestate\Api\Connector');
-        $wsuApi->shouldReceive('sendRequest')->with('cms.promotions.listing', Mockery::type('array'))->once()->andReturn($return);
-
-        // Get the promos
-        $promos = app('App\Repositories\PromoRepository', ['wsuApi' => $wsuApi])->getHomepagePromos();
-
-        $this->assertTrue(is_array($promos));
     }
 }
